code cleaning, fixed removing student/employee from employee/student endpoint

+++ war/src/main/java/pl/poznan/uam/Controllers/EmployeeController.java
@@ -52,6 +52,7 @@ public class EmployeeController {
     @Consumes("application/json; charset=UTF-8")
     @Produces("application/json; charset=UTF-8")
     public Response removeEmployee(@PathParam("id") long id) {
+        if(personDAO.getPersonById(id).get().getPosition().equals("student")) return Response.status(400).build();
         personDAO.remove(id);
         return Response.status(204).build();
     }
